\name{spphpr}
\encoding{UTF-8}
\alias{spphpr}
\docType{package}
\title{
Spring Phenological Prediction
}
\description{
Predicts the occurrence times (in day of year) of spring phenological events. 
Three methods, including the accumulated degree days (ADD) method, 
the accumulated days transferred to a standardized temperature (ADTS) method, 
and the accumulated developmental progress (ADP) method, were used. 
See Shi et al. (2017a, 2017b) for details.  
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{spphpr}
\packageIndices{spphpr}
}
\note{
We thank Benjamin Altmann, Lei Chen, Linli Deng, Feng Ge, Wen Gu, Liang Guo, Jianguo Huang, 
Cang Hui, Konstanze Lauseker, Gadi V.P. Reddy, Jing Tan, Mei Xiao, Lin Wang, 
and Wangxiang Zhang for their valuable help during the development of this package.
}
\author{
Peijian Shi [aut, cre], Zhenghong Chen [aut], Brady K. Quinn [aut]

Maintainer: Peijian Shi <pjshi@njfu.edu.cn>
}
\references{
Shi, P., Chen, Z., Reddy, G.V.P., Hui, C., Huang, J., Xiao, M. (2017a) Timing of cherry tree blooming: 
Contrasting effects of rising winter low temperatures and early spring temperatures. 
\emph{Agricultural and Forest Meteorology} 240\eqn{-}241, 78\eqn{-}89. \doi{10.1016/j.agrformet.2017.04.001}

Shi, P., Fan, M., Reddy, G.V.P. (2017b) Comparison of thermal performance equations in describing 
temperature-dependent developmental rates of insects: (III) Phenological applications.  
\emph{Annals of the Entomological Society of America} 110, 558\eqn{-}564. \doi{10.1093/aesa/sax063}
}

\keyword{ package }


